Introduce the server
This commit introduces the server that will return the rabbit.

== Design Notes
=== Error return

This commit uses the "%w: %s" fmt style of errors, returning an error
entity that can be asserted in tests (or runtime logic) while hiding the
details of that function.

It is not used here, but it does make things nicer to read.

=== Server

This commit creates the "Server" application in a separate package. The
existing server is still problematic unless it is run with "sudo", as it
requests (by default) port 80. The goal here is to have this also be a
demonstration of the logs.

The server also rejects requests where the user agent is Elmar Fudd.
This is designed to be something that shows people the value of
distributed tracing; notably, that specific requests can be useful.