Add application lifecycle logging
This commit introduces the first "Pillar" of observability: Logging.
Logging, while useful, should generally be limited to things that do not
have a defined "start" and "end", or event processing — rather, they
should be used for state transitions. The vast majority of cases,
they're part of application lifecycle handling.

== Design Notes
=== Moving Execute

This commit moves execute, as it is easier to screenshot with it all in
one line.