Integrate the tracer provider into the server
This commit makes the server receive a tracer provider, and generate a
tracer.

== Design Notes
=== Extending the HTTP server object

The simplest modification from the existing design was simply to create
a new struct that receives a tracer, but that is otherwise equilvelant
to the HTTP server struct.

Given this, it is simply a new struct with the old embedded. The HTTP
server (still) is returned to the client, as the API doesn't need to
break.